requirements.txt
click
Flask
gunicorn
MarkupSafe
Werkzeug
slack_bolt
python-dotenv
apscheduler
pytz
cohere
requests